requirements.txt
ultralytics
opencv-python
numpy
matplotlib
pandas
pyyaml